Fixed bugs
made 404 and index pages more responsive for mobile devices.

added id's to sections and made headers clickable to navigate to specific section
added animation to every section header.

added a dot to indicate which question is opened in the FAQ Section.
Added animation ton the arrow on the FAQ questions.